(fix|cos-759): Fixed infinite loop and floating toolbar position Refactor FloatingLinkToolbar and improve toolbar positioning
Refactored FloatingLinkToolbar by moving link-related logic and presentation into a new LinkComponent. This was done for better code readability and maintainability by separating concerns.

In the same process, the positioning formula was also optimized in the LinkComponent. Changes were made to handle special conditions where the toolbar overflows the parent container. The improved positioning logic prevents the toolbar from extending beyond the parent boundary, improving user experience. The background color of the positioner was also made important for appropriate visibility.

These changes improve the usability of the text editor by providing a better user experience and enhancing the maintainability of the code base.
