+++ packages/apps/spaces/styles/remirror-editor.scss
@@ -3953,5 +3953,5 @@ a {
 }
 
 .remirror-positioner {
-  background: #B4D5FE;
+  background: #B4D5FE !important;
 }
